requirements.txt
# Web Framework
Flask==3.1.0
Flask-SocketIO==5.4.1
Flask-CORS==5.0.0

# WebSocket
python-socketio==5.12.0
eventlet==0.37.0

# Production Server
gunicorn==23.0.0

# Database
psycopg2-binary==2.9.11

# Configuration
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3
